+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [In development] - Unreleased
 
 
+## [1.5.0] - 2022-01-24
+
+### Added
+
+- Option: Use this Discord Webhook for replies as well? (This is a checkbox in the
+  board settings. When checked, the provided Discord webhook will also be used to
+  send a message for every reply in this board to Discord. Chose wisely!)
+  ![Admin Board Options](aa_forum/docs/images/admin-board-options.jpg)
+
+### Changed
+
+- Webhook handling changed to dhooks-lite, so we can make our Discord messages nicer
+
+
 ## [1.4.3] - 2022-01-24
 
 ### Fixed